Request creation, GetManagerStaffTransactions, GetEmployeeTransaction, GetEmployeeInDepartment, refreshToken

+++ RequestsManagementSystem/Dtos/EmployeeDtos/EmployeeDto.cs
@@ -6,9 +6,11 @@ namespace RequestsManagementSystem.Dtos.EmployeeDtos
         public int EmployeeId { get; set; }
         public string EmployeeName { get; set; } = string.Empty;
         public string DepartmentName { get; set; } = string.Empty;
+        // remainding casual leaves
         public int CasualLeaveCount { get; set; }
         public DateOnly DateOfEmployment { get; set; }
         public string ManagerName { get; set; } = string.Empty;
-        public int RegularLeaveCount { get; set; }
+        // remainding regular leaves
+        public float RegularLeaveCount { get; set; }
     }
 }
